+++ pages/moderator/CatalogIndMod.js
@@ -1,0 +1,17 @@
+import { Component } from "react";
+import { SafeAreaView, Text } from "react-native";
+
+class CatalogIndMod extends Component{
+
+    render(){
+
+        return(
+
+            <SafeAreaView>
+                <Text>{this.props.route.params.productId}</Text>
+            </SafeAreaView>
+        )
+    }
+}
+
+export default CatalogIndMod