+++ pages/moderator/CatalogMod.js
@@ -87,7 +87,7 @@ class CatalogMod extends Component {
 
 
   renderItem = ({ id, product_name, product_price, product_image }) => (
-    <Pressable key={id} onLongPress={() => {
+    <Pressable key={id} onPress={()=>{this.props.navigation.navigate('CatalogIndMod',{productId: id})}} onLongPress={() => {
       this.setModalVisible(!this.state.modalVisible)
       this.setItemsId(id)
     }}>